Resolve omission of stream after crash on first run
chance of omitting a stream in the rather rare scenario where it is on its first run (i.e. there is no state bookmark associated to it) and it was syncing and the tap crashed

This appears to be a bug present in mssql (which was the basis for our tap-db2)

The fix replaces queries ALL streams, rather than just `streams_with_state` to build the currently_syncing_stream list.